requirements.txt
streamlit
crewai
crewai-tools
langchain
langchain-openai
langchain-text-splitters
langchain-community
langchain-chroma
python-dotenv
chromadb
pypdf
fpdf2
markdown
